+++ djangocms_references/__init__.py
@@ -1,3 +1,1 @@
 __version__ = "1.4.3"
-
-default_app_config = "djangocms_references.apps.ReferencesConfig"