+++ djangocms_references/compat.py
@@ -1,9 +1,9 @@
-from distutils.version import LooseVersion
+from cms import __version__ as cms_version
 
-import django
+from packaging.version import Version
 
 
-DJANGO_GTE_21 = LooseVersion(django.get_version()) >= LooseVersion("2.1")
+DJANGO_CMS_4_1 = Version(cms_version) >= Version('4.1')
 
 
 def is_versioning_installed():