+++ djangocms_references/test_utils/nested_references_app/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = "djangocms_references.test_utils.nested_references_app.apps.NestedReferencesAppConfig"